Change integers to proper std::uintX_t.

+++ include/array_data_encoder.h
@@ -42,7 +42,7 @@ public:
 
     fec_parameters get_fec_parameters(void) const;
 
-    array_source_block_encoder get_source_block(int sbn) const;
+    array_source_block_encoder get_source_block(uint16_t sbn) const;
 
     std::vector<array_source_block_encoder> get_source_block_iterable(void) const;
 
@@ -50,11 +50,11 @@ public:
 
     int get_data_offset(void) const;
 
-    int symbol_size(void) const;
+    uint16_t symbol_size(void) const;
 
     uint64_t data_length() const;
 
-    int number_of_source_blocks(void) const;
+    uint8_t number_of_source_blocks(void) const;
 
     array_data_encoder(std::vector<uint8_t> array, int offset,
                        fec_parameters fecParams)
@@ -64,7 +64,8 @@ public:
 
 private:
     std::vector<array_source_block_encoder> get_source_block_encoders(void);
-    array_source_block_encoder get_encoder(int offset, int source_block_num, int K);
+    array_source_block_encoder get_encoder(int offset, uint16_t source_block_num,
+                                           uint16_t K);
 
     const std::vector<uint8_t> m_array;
     const int m_offset;
